refactor(studypal): move cartesia-studypal into studypal
