--- cartesia-studypal/env.example
@@ -1,3 +1,0 @@
-DAILY_API_KEY= # Create here: https://dashboard.daily.co/developers 
-OPENAI_API_KEY= # Create here: https://platform.openai.com/docs/overview
-CARTESIA_API_KEY= # Create here: https://play.cartesia.ai/keys
