--- cartesia-studypal/pyproject.toml
@@ -1,22 +1,0 @@
-[project]
-name = "cartesia-studypal"
-version = "0.1.0"
-description = "Cartesia study pal example"
-requires-python = ">=3.11,<4"
-dependencies = [
-    "pipecat-ai[anthropic,google,daily,webrtc,websocket,cartesia,openai,runner]>=1.3.0,<2",
-    "pipecatcloud>=0.4.4,<1",
-    "pypdf>=6,<7",
-    "tiktoken>=0.13.0,<1"
-]
-
-[dependency-groups]
-dev = [
-    "pyright>=1.1.404,<2",
-    "ruff>=0.12.11,<1",
-]
-
-[tool.ruff]
-line-length = 100
-[tool.ruff.lint]
-select = ["I"]
