+++ studypal/pyproject.toml
@@ -1,10 +1,10 @@
 [project]
 name = "studypal"
-version = "0.1.0"
+version = "0.2.0"
 description = "Studypal example"
 requires-python = ">=3.11"
 dependencies = [
-    "pipecat-ai[daily,webrtc,websocket,cartesia,openai,silero,deepgram,runner]>=1.3.0",
+    "pipecat-ai[daily,webrtc,websocket,cartesia,openai,silero,runner]>=1.3.0",
     "pipecatcloud>=0.7.1",
     "pypdf>=4.3.1",
     "tiktoken>=0.12.0"
